# backend/docs/swagger.yaml
basePath: /
definitions:
...
  web-hosting_internal_database_entities.Role:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  web-hosting_internal_modules_role_dto.RoleCreateRequest:
    properties:
      role_name:
        example: raja-sawit // required
        type: string
    required:
    - role_name
    type: object
  web-hosting_internal_modules_role_dto.RoleUpdateRequest:
    properties:
      role_name:
        example: raja-solo // required
        type: string
    required:
    - role_name
    type: object
...
  web-hosting_internal_package_swagger.ErrCreateRoleFailed:
    properties:
      error:
        example: role already exists
        type: string
      message:
        example: failed to create role
        type: string
      path:
        example: /api/super/role
        type: string
      success:
        example: false
        type: boolean
    type: object
  web-hosting_internal_package_swagger.ErrCreateRoleInternalServer:
    properties:
      error:
        example: Internal Error
        type: string
      message:
        example: failed to create role
        type: string
      path:
        example: /api/super/role
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  web-hosting_internal_package_swagger.ErrDeleteRoleFailed:
    properties:
      error:
        example: role not found
        type: string
      message:
        example: failed to delete role
        type: string
      path:
        example: /api/super/role/mahasiswa
        type: string
      success:
        example: false
        type: boolean
    type: object
  web-hosting_internal_package_swagger.ErrDeleteRoleInternalServer:
    properties:
      error:
        example: Internal Error
        type: string
      message:
        example: failed to delete role
        type: string
      path:
        example: /api/super/role/mahasiswa
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  web-hosting_internal_package_swagger.ErrForbiddenAccess:
    properties:
      error:
        example: Forbidden
        type: string
      message:
        example: Role anda tidak diizinkan
        type: string
      path:
        example: /api/super/role
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  web-hosting_internal_package_swagger.ErrGetRoleFailed:
    properties:
      error:
        example: Internal Error
        type: string
      message:
        example: failed to get role
        type: string
      path:
        example: /api/role
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  web-hosting_internal_package_swagger.ErrUnauthorizedInvalidToken:
    properties:
      error:
        example: invalid token
        type: string
      message:
        example: failed_auth
        type: string
      path:
        example: /api/auth/logout
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  web-hosting_internal_package_swagger.ErrUpdateRoleFailed:
    properties:
      error:
        example: role not found
        type: string
      message:
        example: failed to update role
        type: string
      path:
        example: /api/super/role/mahasiswa
        type: string
      success:
        example: false
        type: boolean
    type: object
  web-hosting_internal_package_swagger.ErrUpdateRoleInternalServer:
    properties:
      error:
        example: Internal Error
        type: string
      message:
        example: failed to update role
        type: string
      path:
        example: /api/super/role/mahasiswa
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  web-hosting_internal_package_utils.Response-any-any:
    properties:
      data: {}
      error: {}
      message:
        example: Operation successful
        type: string
      path:
        example: /api/resource
        type: string
      success:
        type: boolean
    type: object
  web-hosting_internal_package_utils.Response-array_web-hosting_internal_database_entities_Role-any:
    properties:
      data:
        items:
          $ref: '#/definitions/web-hosting_internal_database_entities.Role'
        type: array
      error: {}
      message:
        example: Operation successful
        type: string
      path:
        example: /api/resource
        type: string
      success:
        type: boolean
    type: object
...
  web-hosting_internal_package_utils.Response-web-hosting_internal_database_entities_Role-any:
    properties:
      data:
        $ref: '#/definitions/web-hosting_internal_database_entities.Role'
      error: {}
      message:
        example: Operation successful
        type: string
      path:
        example: /api/resource
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: |-
    **BARANG SIAPA YANG UPDATE/HAPUS DATA SEMBARANGAN DAPAT KECUPAN DARI HAFIZ**.
    **BUAT DATA BARU JIKA INGIN HAPUS/UPDATE**.

    **DUMMY DATA LINK:**
    **https://github.com/Kar-Su/pbl-project-simpadu/tree/main/backend/internal/database/seeders/json**

    **TUTORIAL LINK:***
    **https://github.com/Kar-Su/pbl-project-simpadu**
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: TIM 1 API
  version: "1.0"
paths:
...
  /api/roles:
    get:
      description: |-
        Mengambil seluruh daftar role yang tersedia di sistem.

        **Akses:** Super Admin, Admin Akademik, Admin Keuangan, Admin Mahasiswa.

        **Error yang mungkin terjadi:**
        - `401` Authorization header tidak ada -> `message: "failed_auth", error: "Authorization header missing"`
        - `401` Format header salah (bukan "Bearer ...") -> `message: "failed_auth", error: "invalid authentication header"`
        - `401` Token JWT tidak valid atau kedaluwarsa -> `message: "failed_auth", error: "invalid token"`
        - `403` Role user tidak memiliki akses -> `message: "Role anda tidak diizinkan", error: "Forbidden"`
        - `500` Kesalahan internal server -> `message: "failed to get role", error: "Internal Error"`
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web-hosting_internal_package_utils.Response-array_web-hosting_internal_database_entities_Role-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrUnauthorizedInvalidToken'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrForbiddenAccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrGetRoleFailed'
      security:
      - ApiKeyAuth: []
      summary: Ambil Semua Role
      tags:
      - role
    post:
      consumes:
      - application/json
      description: |-
        Menambahkan role baru ke dalam sistem.

        **Akses:** Super Admin.

        **Error yang mungkin terjadi:**
        - `400` Body tidak valid / field wajib kosong -> `message: "failed to get request", error: "Key: 'RoleName' Error:..."`
        - `400` Role dengan nama tersebut sudah ada -> `message: "failed to create role", error: "role already exists"`
        - `401` Authorization header tidak ada -> `message: "failed_auth", error: "Authorization header missing"`
        - `401` Format header salah (bukan "Bearer ...") -> `message: "failed_auth", error: "invalid authentication header"`
        - `401` Token JWT tidak valid atau kedaluwarsa -> `message: "failed_auth", error: "invalid token"`
        - `403` Role user tidak memiliki akses -> `message: "Role anda tidak diizinkan", error: "Forbidden"`
        - `500` Kesalahan internal server -> `message: "failed to create role", error: "Internal Error"`
      parameters:
      - description: Payload Create Role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/web-hosting_internal_modules_role_dto.RoleCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/web-hosting_internal_package_utils.Response-web-hosting_internal_database_entities_Role-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrCreateRoleFailed'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrUnauthorizedInvalidToken'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrForbiddenAccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrCreateRoleInternalServer'
      security:
      - ApiKeyAuth: []
      summary: Buat Role Baru
      tags:
      - role
  /api/roles/{role_name}:
    delete:
      consumes:
      - application/json
      description: |-
        Menghapus role dari sistem secara permanen berdasarkan nama role.

        **Akses:** Khusus Super Admin.

        **Error yang mungkin terjadi:**
        - `400` Parameter URI tidak valid -> `message: "failed to validate role uri", error: "Key: 'RoleName' Error:..."`
        - `400` Role dengan nama tersebut tidak ditemukan -> `message: "failed to delete role", error: "role not found"`
        - `401` Authorization header tidak ada -> `message: "failed_auth", error: "Authorization header missing"`
        - `401` Format header salah (bukan "Bearer ...") -> `message: "failed_auth", error: "invalid authentication header"`
        - `401` Token JWT tidak valid atau kedaluwarsa -> `message: "failed_auth", error: "invalid token"`
        - `403` Role user tidak memiliki akses -> `message: "Role anda tidak diizinkan", error: "Forbidden"`
        - `500` Kesalahan internal server -> `message: "failed to delete role", error: "Internal Error"`
      parameters:
      - description: Nama Role yang Akan Dihapus
        example: mahasiswa
        in: path
        name: role_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web-hosting_internal_package_utils.Response-any-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrDeleteRoleFailed'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrUnauthorizedInvalidToken'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrForbiddenAccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrDeleteRoleInternalServer'
      security:
      - ApiKeyAuth: []
      summary: Hapus Role
      tags:
      - role
    put:
      consumes:
      - application/json
      description: |-
        Mengubah nama role berdasarkan nama role yang sudah ada.

        **Akses:** Super Admin.

        **Error yang mungkin terjadi:**
        - `400` Parameter URI tidak valid -> `message: "failed to validate role uri", error: "Key: 'RoleName' Error:..."`
        - `400` Body tidak valid / field wajib kosong -> `message: "failed to get request", error: "Key: 'RoleName' Error:..."`
        - `400` Role dengan nama tersebut tidak ditemukan -> `message: "failed to update role", error: "role not found"`
        - `401` Authorization header tidak ada -> `message: "failed_auth", error: "Authorization header missing"`
        - `401` Format header salah (bukan "Bearer ...") -> `message: "failed_auth", error: "invalid authentication header"`
        - `401` Token JWT tidak valid atau kedaluwarsa -> `message: "failed_auth", error: "invalid token"`
        - `403` Role user tidak memiliki akses -> `message: "Role anda tidak diizinkan", error: "Forbidden"`
        - `500` Kesalahan internal server -> `message: "failed to update role", error: "Internal Error"`
      parameters:
      - description: Nama Role yang Akan Diubah
        example: mahasiswa
        in: path
        name: role_name
        required: true
        type: string
      - description: Payload Update Role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/web-hosting_internal_modules_role_dto.RoleUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web-hosting_internal_package_utils.Response-web-hosting_internal_database_entities_Role-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrUpdateRoleFailed'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrUnauthorizedInvalidToken'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrForbiddenAccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrUpdateRoleInternalServer'
      security:
      - ApiKeyAuth: []
      summary: Update Role
      tags:
      - role
...
securityDefinitions:
  ApiKeyAuth:
    description: 'Format: Bearer <access token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"